+++ src/ImageResizer.Plugins.AutoCrop/Models/AutoCropState.cs
@@ -1,0 +1,22 @@
+﻿using ImageResizer.Plugins.AutoCrop.Analyzers;
+using System.Drawing;
+
+namespace ImageResizer.Plugins.AutoCrop.Models
+{
+    public class AutoCropState
+    {
+        public readonly Rectangle Bounds;
+        public readonly Color BorderColor;
+        public readonly Rectangle OriginalDimensions;
+        public Rectangle TargetDimensions;
+        public Size Padding;
+        public float Scale;
+
+        public AutoCropState(BoundsAnalyzer analyzer, Bitmap bitmap)
+        {
+            Bounds = analyzer.BoundingBox;
+            BorderColor = analyzer.BorderAnalysis.BackgroundColor;
+            OriginalDimensions = new Rectangle(0, 0, bitmap.Width, bitmap.Height);
+        }
+    }
+}