+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4.0]
+
+### Added
+- Improved image analysis (combined bucket analysis and color thresholding)
+- Improved layout capabilities, will now add padding to images that don't have enough whitespace.
+
 ## [1.3.0]
 
 ### Added